sfp: correct setting mantissa to use "=" and not "+="
When setting the mantissa for an rx_pwr calibration constant it is using
"+=" instead of "+". This means that initially it could take a value off
the stack, and then each subsequent rc_pwr it will add in the mantissa
of the previous one. This has been corrected by setting it using "="
instead of "+=".

VRVDR-56385