sfp: correct the rx_pwr externally calibrated calculation
The externally calibrated calculation for rx_pwr from the SFF-8472
standard says that the multipliers for the raw power values for Rx_PWR(2),
Rx_PWR(3), and Rx_PWR(4), should be the raw rx_pwr number to the power
of 2, 3, and 4 respectively. However, the code instead multiplyes by
the raw power values times 10 to the power of 2, 3, and 4.
The formula is:

Rx_PWR = Rx_PWR(4) x RawRxPWR^4 +
         Rx_PWR(3) x RawRxPWR^3 +
         Rx_PWR(2) x RawRxPWR^2 +
         Rx_PWR(1) x RawRxPWR +
         Rx_PWR(0)

This update corrects this.

VRVDR-56385