sfp: ensure calibrated values are within allowed range
The code has been updated to ensure that externally calibrated values
have the same range of values as internally calibrated. This is so that
returned values will not be outside the supported ranges as given in
the YANG RPC definitions.

Each type of value (temperature, power, current bias, and voltage) for
internally calibrated is stored as a 16-bit number (unsigned, except for
temperature which is signed).  As the externally calibrated value should
give the same range, the updated code ensures they are within the range
of 16-bit values.  for internally calibrated.

VRVDR-56385