json: add functions to write floating to a precision
Added functions jsonw_float_prec_field() and jsonw_float_prec() to write
a floating point number in JSON to a chosen precision.  This is to allow
the numbers returned to match the representation used in YANG. Also the
number is written as a string, for compatibility with YANG, which cannot
handle floating point numbers with a decimal point (which would end up
removing all digits after the decimal point).

VRVDR-56385