sfp: ensure the JSON values have the required precision
The YANG code expects 2 digits of precision after the decimal point.
If it gets more digits when expected, then it gives error: "Invalid
decimal64: Number of fractional digits in input string is greater
than int parameter".

The current code can return more digits, but instead of causing this
error, the digits after the decimal point are removed due to the way
the YANG code handles JSON floating point numbers, when they are not quoted
as a string. This results in values displayed rounded down to whole
numbers, e.g. a voltage of "3.75" is printed as "3".

To resolve this use the newly added JSON function
(jsonw_float_prec_field()) which (a) can print a number to a given
precision, and (b) returns it as a string, for compatibility with YANG.

As power gets converted by the RPC script from mW to dbm, it did not
need the floating point JSON code updated.

VRVDR-56385