sfp: add calibration to threshold values
The functions that calculate and output in JSON the threshold values
for alarms and warnings are not passed in the calibration constants,
and so were not calibrated when external calibration is used by the SFP
transceiver. According to 9.4 of SFF-8472 they should be calibrated,
and so the code has been updated to pass the calibration constants in,
so they are calibrated.

VRVDR-56385